% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigMSCdailyToObs.R
\name{bigMSCdailyToObs}
\alias{bigMSCdailyToObs}
\title{Reads large MSC files of daily values}
\usage{
bigMSCdailyToObs(infile, quiet = TRUE, logfile = "")
}
\arguments{
\item{infile}{Required. Name of the file to be read.}

\item{quiet}{Optional.Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional.  Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns TRUE. If unsuccessful, returns the value FALSE.
}
\description{
Reads large MSC files holding daily values of several variables at several sites and exports an hourly CRHM \code{obs} data file for each site. The obs files are of the form \option{<sitenumber>_daily.obs'}
}
\examples{
\dontrun{
bigMSCdailyToObs('GRPextractor_PHW_Bad_Lake_dlyv2_21032015_155647.txt', quiet=FALSE)}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{bigMSChourlyToObs}}
}

