% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigMSChourlyToObs.R
\name{bigMSChourlyToObs}
\alias{bigMSChourlyToObs}
\title{Reads large MSC files of hourly values}
\usage{
bigMSChourlyToObs(infile, timezone = "", quiet = TRUE, logfile = "")
}
\arguments{
\item{infile}{Required. Name of the file to be read.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{quiet}{Optional. ptional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional.  Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns TRUE. If unsuccessful, returns the value FALSE.
}
\description{
Reads large MSC files holding hourly values of several variables at several sites and exports an hourly CRHM \code{obs} data file for each site. The obs files are of the form \option{<siten umber>_hourly.obs'}
}
\examples{
\dontrun{
bigMSChourlyToObs('GRPextractor_PHW_Bad_Lake_hly_21032015_152943.txt', timezone='CST', quiet=FALSE)}
}
\author{
Kevin Shook
#' @seealso  \code{\link{bigMSCdailyToObs}}
}
