% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/downloadMSCobs.R
\name{downloadMSCobs}
\alias{downloadMSCobs}
\title{Downloads monthly and daily MSC data and creates CRHM .obs files.}
\usage{
downloadMSCobs(station.name = "", station.number = "", startyear = 1900,
  endyear = 2000, timezone = "", quiet = TRUE, logfile = "MSCr.log")
}
\arguments{
\item{station.name}{Name of the station as a text string. This is used to create a directory that will hold the downloaded files and the created .obs files. It is also the basis for the names of the .obs files: <station.name>Hourly.obs, <station.name>DailyTemps.obs, and <station.name>DailyPrecips.obs.}

\item{station.number}{Required,}

\item{startyear}{Optional. First year for downloading. Default is 1900.}

\item{endyear}{Optional Last year for downloading. Default is 2000.}

\item{timezone}{Optional. The name of the timezone of the data as a character string. This should be YOUR timezone. The default value is "", which is your timezone. You may find a difference in the seconds between using "" and the your timezone. Note that the timezone code is specific to your OS. Under Windows, the code for Central Standard Time is 'America/Regina'. Under Linux, it is 'CST'.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. Note that the default is FALSE. If you are calling this function in an R script, you will usually leave quiet=TRUE (i.e. the default). If you are working interactively, you will probably want to set quiet=FALSE.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used. Default is 'MSCr.log'}
}
\value{
If successful, returns TRUE. If unsuccessful, returns the value FALSE.
}
\description{
Downloads MSC hourly and daily data, one month at a time. The downloaded data are stored as .csv files, which are erased after the function terminates. Information about the .obs files is displayed on the screen and is also written to log files in the working directory. The obs files are: 1) hourly t, rh, and u  2) daily tmin, tmax, tmean and 3) daily precipitation. Because Environment Canada's web data is not consistent in any way, it is possble that using this function will generate warning messages.
}
\examples{
downloadMSCobs('Vegreville', 1977, 1995, 1996)
}
\author{
Kevin Shook
}
\references{
The code for downloading MSC data is taken from \url{http://www.fromthebottomoftheheap.net/2015/01/14/harvesting-canadian-climate-data}. Some modifications were made to remove bad characters, and to download daily values.
}

