% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMSCstations.R
\name{downloadMSCstations}
\alias{downloadMSCstations}
\title{Reads current set of MSC stations from server}
\usage{
downloadMSCstations(stationURL = "", outfile = "", quiet = TRUE,
  logfile = "")
}
\arguments{
\item{stationURL}{Optional. URL of file containing stations. If not specified then the file at "client_climate@ftp.tor.ec.gc.ca" will be used.}

\item{outfile}{Optional. If specified, then the stations will be written to a comma delimited file.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}. Note that setting \code{quiet=FALSE} shows the downloading progress bar.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns a \pkg{CRHMr} data frame containing the station names and their metadata.
}
\description{
Reads current set of MSC stations from server
}
\note{
The station data frame contains the variables: Name, Province, Climate ID, Station ID, WMO ID, TC ID, Latitude (Decimal Degrees), Longitude (Decimal Degrees), Latitude, Longitude, Elevation (m), First Year, Last Year, HLY First Year, HLY Last Year, DLY First Year, DLY Last Year, MLY First Year, MLY Last Year.
}
\examples{
\dontrun{
MSCstations <- downloadMSCstations(quiet=FALSE)}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{downloadMSCobs}}
}

