% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMSCDailyTtoObs.R
\name{oldMSCdailyTtoObs}
\alias{oldMSCdailyTtoObs}
\title{Creates an obs file from all MSC old-style daily temperature files in a directory}
\usage{
oldMSCdailyTtoObs(directory = ".", filespec = "A1*", timezone = "")
}
\arguments{
\item{directory}{Optional. Directory containing all temperature files. The default is the current directory.}

\item{filespec}{Optional. File specification for all files. Default is \option{A1*}.}

\item{timezone}{Required.  Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}
}
\value{
If successful returns \code{TRUE}. If unsuccessful returns \code{FALSE}.
}
\description{
Reads all files of old-style MSC daily soil, min and max air temp. data from a specified directory and assembles them. The values are organized in columns, 1 year per file. Writes a obs file called \option{xx_tminmax.obs}, where \option{xx} is the first 2 characters of \option{filespec}.
}
\examples{
\dontrun{
oldMSCHourlyUtoObs(timezone='Etc/GMT+7')
}
}
