% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAESdailyP.R
\name{readAESdailyP}
\alias{readAESdailyP}
\title{Reads individual AES files of daily precipitation data in a directory and creates obs files.}
\usage{
readAESdailyP(directory = ".", filespec = "A0O*", timezone = "")
}
\arguments{
\item{directory}{Optional. Directory containing files. If not specified, defaults to current directory. Note that this is an R path, which uses the \code{'/'} symbol on ALL operating systems.}

\item{filespec}{Optional. File specification (including wildcards) of the precipitation data. Default is \option{'A0O*'}.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}
}
\description{
Reads very old Atmospheric Environment Service (AES) files. All of the files specified are assembled to create a \code{CRHM} observation file named \option{ppt.obs} in the same directory as the data files. It is assumed that the files only contain days with recorded precipitation values, so days with zero values are inserted.
}
\examples{
\dontrun{
readAESdailyP('./HistoricalData', timezone='etc/GMT+6')
}
}
\seealso{
\code{\link{readAEShourlyRH}} \code{\link{readAEShourlyWind}} \code{\link{readAEShourlyT}} code{\link{readAESdailyTminTmax}}
}
\author{
Kevin Shook
}
