% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAHCCDdailyTemps.R
\name{readAHCCDdailyTemps}
\alias{readAHCCDdailyTemps}
\title{Reads AHCCD daily temperature values}
\usage{
readAHCCDdailyTemps(infile)
}
\arguments{
\item{infile}{Required. Name of the file to be read.}
}
\value{
If successful, returns the values in a dataframe, consisting of the date, the value and the data code. If unsuccessful, returns the value FALSE.
}
\description{
This program reads Adjusted and Homogenized Canadian Climate Data (AHCCD) data in a month x day data file of min max or mean air temperatures.
}
\note{
The AHCCD are often used for statistical analysis, so this function returns a data quality code, and does not create a CRHM obs files.
}
\examples{
\dontrun{
stoon.tmax <- readAHCCDdailyTemps('dx40657120.txt')}
}
\author{
Kevin Shook
}
\references{
Monthly AHCCD data are available from \url{http://www.ec.gc.ca/dccha-ahccd}. Daily values must be requested. Any use of the temperature data must cite \cite{Vincent, L. A., X. L. Wang, E. J. Milewska, H. Wan, F. Yang, and V. Swail , 2012. A second generation of homogenized Canadian monthly surface air temperature for climate trend analysis, J. Geophys. Res., 117, D18110, doi:10.1029/2012JD017859.}
}
\seealso{
\code{\link{readAHCCDdailyTemps}} \code{\link{readAHCCDmonthlyPrecips}}
}

