% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAHCCDmonthlyPrecips.R
\name{readAHCCDmonthlyPrecips}
\alias{readAHCCDmonthlyPrecips}
\title{Reads AHCCD monthly precipitation values.}
\usage{
readAHCCDmonthlyPrecips(infile)
}
\arguments{
\item{infile}{Required. Name of the file to be read.}
}
\value{
If successful, returns the values in a dataframe, consisting of the date, the value and the data code. If unsuccessful, returns the value FALSE.
}
\description{
This program reads Adjusted and Homogenized Canadian Climate Data (AHCCD) data in a month x day data file of rainfall, snowfall or total precipitation.
}
\note{
The AHCCD are often used for statistical analysis, so this function returns a data quality code, and does not create a CRHM obs file.
}
\examples{
\dontrun{
stoon.monthly.total <- readAHCCDmonthlyPrecips('mt4057120.txt')}
}
\references{
Monthly AHCCD data are available from \url{http://www.ec.gc.ca/dccha-ahccd}. Any use of the precipitation data must cite \cite{Mekis, E and L.A. Vincent, 2011: An overview of the second generation adjusted daily precipitation dataset for trend analysis in Canada. Atmosphere-Ocean, 49 (2), 163-177.}
}
\seealso{
\code{\link{readAHCCDdailyTemps}} \code{\link{readAHCCDdailyPrecips}}
}
\author{
Kevin Shook
}
