% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDatamartHourlyPrecip.R
\name{readDatamartHourlyPrecip}
\alias{readDatamartHourlyPrecip}
\title{Reads hourly Datamart-format precipitation data}
\usage{
readDatamartHourlyPrecip(infile, timezone = "")
}
\arguments{
\item{infile}{Required. File to read precipitation from.}

\item{timezone}{Optional. Timezone of the orginal data. If not specified, your timezone will be used.}
}
\value{
Returns a data frame consisting of the variables 'datetime', 'precip' and 'code'. The precipitation is in mm. The code is the data quality code from the original dataset. Missing values coded as -9999 in the original file are set to \code{NA_real_}.
}
\description{
Reads hourly precipitation values from a file in the MSC Datamart format (day x hour).
}
\examples{
\dontrun{
precip <- readDatamartHourlyPrecip('L1012710.262')}
}

