% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNOAAprecip.R
\name{readNOAAprecip}
\alias{readNOAAprecip}
\title{Read NOAA file of precipitation data}
\usage{
readNOAAprecip(NOAAfile = "", outLoc = "./", timezone = "",
  quiet = TRUE, logfile = "")
}
\arguments{
\item{NOAAfile}{Required. File of NOAA precipitation data. Contains PRECIP and/or SNOWFALL variables. 
May contain data for any number of locations.}

\item{outLoc}{Optional. Location for output files. Each variable for each location will be written to a separate \pkg{CRHMr} obs file.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, 
but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
Read NOAA file of precipitation data
}
\examples{
\dontrun{result <- readNOAAprecip(NOAAfile = "4553527558696dat.txt", timezone = "MST")
}
}
