% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRadarPrecip.R
\name{readRadarPrecip}
\alias{readRadarPrecip}
\title{Reads MSC radar precipitation.}
\usage{
readRadarPrecip(radarFile = "", outputFormat = "grid")
}
\arguments{
\item{radarFile}{Required. File containing precipitation data}

\item{outputFormat}{Optional. Either \option{grid} (the default) or \option{XYZ}.}
}
\value{
If successful, returns radar data in specified format, which consists of a list containing the header information and either grided (matrix) or XYZ (dataframe) values. The precipitation vaues are in mm/hr. If unsuccessful, returns \code{FALSE}.
}
\description{
Reads in file containing MSC radar precipitation data and adds the missing (zero) values. The data consist of a header, followed by a single line containing the word \option{DATA}  and all of the values, which are comma-delimited.
}
\examples{
\dontrun{
readRadarPrecip('junk.num')}
}

