% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWebsiteDailyCSV.R
\name{readWebsiteDailyCSV}
\alias{readWebsiteDailyCSV}
\title{Reads file of daily values produced by MSC website}
\usage{
readWebsiteDailyCSV(websiteCSV, metaData = TRUE)
}
\arguments{
\item{websiteCSV}{Required. A .csv file of daily values as produced by \url{http://climate.weather.gc.ca/historical_data/search_historic_data_e.html}.}

\item{metaData}{Optional. If \code{TRUE} (the default) the meta data are returned.}
}
\value{
If \code{metadata = TRUE}, a list of data frames with the names \code{header}
 (station info), \code{legend} (meaning of the flags) and \code{values}
 daily values are returned. If \code{metadata = FALSE}, then only
 a data frame of the values is returned. In both cases, the values are a time
 series - the first column is a standard {R} \code{Date}.
}
\description{
Reads file of daily values produced by MSC website
}
\examples{
\dontrun{vals <- readWebsiteDailyCSV("eng-daily-01012016-12312016.csv", FALSE)}
}
\author{
Kevin Shook
}
