% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMSCobs.R
\name{downloadMSCobs}
\alias{downloadMSCobs}
\title{Downloads daily and hourly MSC data and creates CRHM .obs files}
\usage{
downloadMSCobs(station.name = "", station.number = "", startyear = 1900,
  endyear = 2000, timezone = "", quiet = TRUE, logfile = "")
}
\arguments{
\item{station.name}{Name of the station as a text string. This is used to create a directory that will hold the downloaded files and the created .obs files. It is also the basis for the names of the .obs files: \option{<station.name>Hourly.obs}, \option{<station.name>DailyTemps.obs}, and \option{<station.name>DailyPrecips.obs}.}

\item{station.number}{Required,}

\item{startyear}{Optional. First year for downloading. Default is \code{1900}.}

\item{endyear}{Optional. Last year for downloading. Default is \code{2000}.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional.  Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns TRUE. If unsuccessful, returns the value FALSE.
}
\description{
Downloads MSC hourly and daily data, one month at a time. The downloaded data are stored as .csv files, which are erased after the function terminates. Information about the .obs files is displayed on the screen and is also written to log files in the working directory. The obs files are: 1) hourly \code{t}, \code{rh}, and \code{u}  2) daily \code{tmin}, \code{tmax}, \code{tmean} and 3) daily precipitation. Because Environment Canada's web data is not consistent in any way, it is possble that using this function will generate warning messages.
}
\examples{
\dontrun{
downloadMSCobs('Vegreville', 1977, 1995, 1996, 'MST')}

}
\references{
The code for downloading MSC data is taken from \url{http://www.fromthebottomoftheheap.net/2015/01/14/harvesting-canadian-climate-data}. Some modifications were made to remove bad characters, and to download daily values.
}
\author{
Kevin Shook
}
